+++ tst/unsupported_input.t
@@ -7,3 +7,9 @@ Complain about non-supported input:
   $ show doesnt_exist
   doesnt_exist: Unsupported input
   [1]
+
+Deactivate non-supported plugin automatically:
+
+  $ PATH= VALGRIND_LEAK_CHECK= show /
+  /: Unsupported input
+  [1]