+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # 0.12.0 - UNRELEASED #
 
 - Rename plugin "pipe" to "stdin" since it only deals with handling standard input.
-- Fix handling of many plugin options.
 - Reuse input buffer for performance (near-equal to plain `ls` now).
+- Fix handling of a large number of plugin options in a single command line call.
 
 # 0.11.0 - 2017-05-03 #