+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - Deactivate plugins automatically when unavailable.
 - Rename plugin "pipe" to "stdin" since it only deals with handling standard input.
 - Fix handling of a large number of plugin options in a single command line call.
-- Reuse input buffer for performance.
+- Fix plugin "pager" to handle SIGINT and terminal resizing correctly.
 
 # 0.11.0 - 2017-05-03 #